Refactor tile import scripts to use common library.
