+++ xc7/utils/prjxray_arch_import.py
@@ -249,16 +249,17 @@ PREFIX_REQUIRED = {
 
 
 def make_prefix(site_name, x, y):
-    k, _ = site_name.split('_')
+    """ Make tile FASM prefix for a given site. """
+    site_type, _ = site_name.split('_')
 
-    prefix_required = PREFIX_REQUIRED[k]
+    prefix_required = PREFIX_REQUIRED[site_type]
     if prefix_required[0] == 'Y':
-        return k, '{}_Y{}'.format(k, y % prefix_required[1])
+        return site_type, '{}_Y{}'.format(site_type, y % prefix_required[1])
     elif prefix_required[0] == 'XY':
         mod_x, mod_y = prefix_required[1]
-        return k, '{}_X{}Y{}'.format(k, x % mod_x, y % mod_y)
+        return site_type, '{}_X{}Y{}'.format(site_type, x % mod_x, y % mod_y)
     else:
-        assert False, k
+        assert False, (site_type, prefix_required)
 
 
 def get_site_prefixes(conn, tile_pkey):
@@ -289,13 +290,20 @@ AND
   """, (tile_pkey, tile_pkey, tile_pkey)
     )
     for site_name, x, y in cur:
-        k, prefix = make_prefix(site_name, x, y)
-        site_prefixes[k] = prefix
+        site_type, prefix = make_prefix(site_name, x, y)
+        assert site_type not in site_prefixes
+        site_prefixes[site_type] = prefix
 
     return site_prefixes
 
 
-def create_capacity_prefix(c, tile_prefix, tile_pkey):
+def create_capacity_prefix(c, tile_prefix, tile_pkey, tile_capacity):
+    """ Create FASM prefixes for all sites located within specified tile.
+
+    This function should be invoke when the relevant tile has capacity > 1
+    and therefore there should be prefixes for each instance.
+
+    """
     c.execute(
         """
 SELECT site_instance.name, site_instance.x_coord, site_instance.y_coord
@@ -314,17 +322,18 @@ ORDER BY site_type.name, site_instance.x_coord, site_instance.y_coord;""",
         (tile_pkey, )
     )
 
-    def inner():
-        for site_name, x, y in c:
-            k, prefix = make_prefix(site_name, x, y)
-            yield '{}.{}.{}'.format(tile_prefix, k, prefix)
+    prefixes = []
+    for site_name, x, y in c:
+        site_type, prefix = make_prefix(site_name, x, y)
+        prefixes.append('{}.{}.{}'.format(tile_prefix, site_type, prefix))
 
-    return " ".join(inner())
+    assert len(prefixes
+               ) == tile_capacity, (tile_pkey, tile_capacity, len(prefixes))
+
+    return " ".join(prefixes)
 
 
-def get_fasm_tile_prefix(
-        conn, g, tile_pkey, site_as_tile_pkey, tile_has_capacity
-):
+def get_fasm_tile_prefix(conn, g, tile_pkey, site_as_tile_pkey, tile_capacity):
     """ Returns FASM prefix of specified tile. """
     c = conn.cursor()
 
@@ -363,7 +372,7 @@ WHERE
             tile_type_map[tile_type] = tilename
 
     if len(tile_type_map) > 1:
-        assert not tile_has_capacity
+        assert tile_capacity == 1
         tile_type_map.update(get_site_prefixes(conn, tile_pkey))
         return lambda single_xml: attach_multiple_prefixes_to_tile(
             single_xml, tile_type_map
@@ -376,7 +385,7 @@ WHERE
         # If this tile is site_as_tile, add an additional prefix of the site
         # that is embedded in the tile.
         if site_as_tile_pkey is not None:
-            assert not tile_has_capacity
+            assert tile_capacity == 1
             c.execute(
                 "SELECT site_pkey FROM site_as_tile WHERE pkey = ?",
                 (site_as_tile_pkey, )
@@ -398,8 +407,10 @@ WHERE
 
             tile_prefix = '{}.{}_X{}'.format(tile_prefix, site_type_name, x)
 
-        if tile_has_capacity:
-            tile_prefix = create_capacity_prefix(c, tile_prefix, tile_pkey)
+        if tile_capacity > 1:
+            tile_prefix = create_capacity_prefix(
+                c, tile_prefix, tile_pkey, tile_capacity
+            )
 
         return lambda single_xml: attach_prefix_to_tile(
             single_xml, tile_prefix
@@ -443,8 +454,7 @@ def attach_multiple_prefixes_to_tile(single_xml, tile_type_map):
 
 
 def get_tiles(
-        conn, g, roi, synth_loc_map, synth_tile_map, tile_types,
-        tile_has_capacity
+        conn, g, roi, synth_loc_map, synth_tile_map, tile_types, tile_capacity
 ):
     """ Yields tiles in grid.
 
@@ -492,7 +502,7 @@ def get_tiles(
         vpr_tile_type = add_vpr_tile_prefix(tile_type)
 
         meta_fun = get_fasm_tile_prefix(
-            conn, g, tile_pkey, site_as_tile_pkey, tile_has_capacity[tile_type]
+            conn, g, tile_pkey, site_as_tile_pkey, tile_capacity[tile_type]
         )
 
         yield vpr_tile_type, grid_x, grid_y, meta_fun
@@ -643,7 +653,7 @@ def main():
         )
 
     tiles_xml = ET.SubElement(arch_xml, 'tiles')
-    tile_has_capacity = {}
+    tile_capacity = {}
     for tile_type in tile_types:
         uri = tile_xml_spec.format(tile_type.lower())
         ET.SubElement(tiles_xml, xi_include, {
@@ -655,10 +665,9 @@ def main():
 
             tile_root = tile_xml.getroot()
             assert tile_root.tag == 'tile'
-            tile_has_capacity[tile_type] = False
-            if 'capacity' in tile_root.attrib and int(
-                    tile_root.attrib['capacity']) > 1:
-                tile_has_capacity[tile_type] = True
+            tile_capacity[tile_type] = 1
+            if 'capacity' in tile_root.attrib:
+                tile_capacity[tile_type] = int(tile_root.attrib['capacity'])
 
     complexblocklist_xml = ET.SubElement(arch_xml, 'complexblocklist')
     for pb_type in pb_types:
@@ -746,7 +755,7 @@ def main():
                 synth_loc_map=synth_loc_map,
                 synth_tile_map=synth_tile_map,
                 tile_types=tile_types,
-                tile_has_capacity=tile_has_capacity,
+                tile_capacity=tile_capacity,
         ):
             single_xml = ET.SubElement(
                 fixed_layout_xml, 'single', {
